requirements.txt
streamlit==1.32.2
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
joblib==1.3.2
pymongo==4.6.2
python-dotenv==1.0.1
shap==0.44.1
matplotlib==3.8.3
seaborn==0.13.2
flask